requirements.txt
streamlit
pandas
numpy
plotly
fpdf2
openai
google-generativeai

openpyxl
